+++ src/FakeXrmEasy.Plugins/EntityLogicalNameConstants.cs
@@ -1,0 +1,25 @@
+﻿namespace FakeXrmEasy.Plugins
+{
+    internal static class EntityLogicalNameConstants
+    {
+        internal const string AppNotification = "appnotification";
+        internal const string BackgroundOperation = "backgroundoperation";
+        internal const string CardStateItem = "cardstateitem";
+        internal const string ComponentVersionNrdDataSource = "componentversionnrddatasource";
+        internal const string ElasticFileAttachment = "elasticfileattachment";
+        internal const string EventExpanderBreadcrumb = "eventexpanderbreadcrumb";
+        internal const string FlowLog = "flowlog";
+        internal const string FlowRun = "flowrun";
+        internal const string MsDynTimelinePin = "msdyn_timelinepin";
+        internal const string NlsqRegistration = "nlsqregistration";
+        internal const string None = "none";
+        internal const string PowerPagesLog = "powerpageslog";
+        internal const string RecentlyUsed = "recentlyused";
+        internal const string SearchResultsCache = "searchresultscache";
+        internal const string SearchTelemetry = "searchtelemetry";
+        internal const string SharedWorkspaceAccessToken = "sharedworkspaceaccesstoken";
+        internal const string SharedWorkspaceNr = "sharedworkspacenr";
+        
+        internal const string Account = "account";
+    }
+}