+++ src/FakeXrmEasy.Plugins/XrmFakedPluginContextProperties.cs
@@ -6,13 +6,16 @@ using FakeXrmEasy.Abstractions.Plugins;
 using Microsoft.Xrm.Sdk;
 #if FAKE_XRM_EASY_9
 using Microsoft.Xrm.Sdk.PluginTelemetry;
+using FakeXrmEasy.Plugins.PluginExecutionContext;
 #endif
 using System;
+
 
 namespace FakeXrmEasy.Plugins 
 {
     /// <summary>
-    /// Implementation to override default plugin context properties
+    /// Implements several interfaces that are needed for plugin execution (i.e. IServiceProvider, IExecutionContext, IPluginExecutionContext)
+    /// from a given IOrganizationService and ITracingService
     /// </summary>
     public class XrmFakedPluginContextProperties : IXrmFakedPluginContextProperties
     {
@@ -60,7 +63,7 @@ namespace FakeXrmEasy.Plugins
             _serviceEndpointNotificationService = A.Fake<IServiceEndpointNotificationService>();
 
 #if FAKE_XRM_EASY_9
-                _entityDataSourceRetrieverService = A.Fake<IEntityDataSourceRetrieverService>();
+            _entityDataSourceRetrieverService = A.Fake<IEntityDataSourceRetrieverService>();
                 A.CallTo(() => _entityDataSourceRetrieverService.RetrieveEntityDataSource())
                     .ReturnsLazily(() => EntityDataSourceRetriever);
 
@@ -117,11 +120,18 @@ namespace FakeXrmEasy.Plugins
                        return GetFakedPluginContext((XrmFakedPluginExecutionContext) plugCtx);
                    }
 
+#if FAKE_XRM_EASY_9
+                   if (t == typeof(IPluginExecutionContext4))
+                   {
+                       return GetFakedPluginContext4((XrmFakedPluginExecutionContext4) plugCtx);
+                   }
+#endif
+                   
                    if (t == typeof(IExecutionContext))
                    {
                        return GetFakedExecutionContext((XrmFakedPluginExecutionContext) plugCtx);
                    }
-
+                   
                    if (t == typeof(IOrganizationServiceFactory))
                    {
                        return _organizationServiceFactory;
@@ -159,14 +169,29 @@ namespace FakeXrmEasy.Plugins
         {
             var context = A.Fake<IPluginExecutionContext>();
 
-            PopulateExecutionContextPropertiesFromFakedContext(context, ctx);
-
-            A.CallTo(() => context.ParentContext).ReturnsLazily(() => ctx.ParentContext);
-            A.CallTo(() => context.Stage).ReturnsLazily(() => ctx.Stage);
+            PopulatePluginExecutionContextPropertiesFromFakedContext(context, ctx);
 
             return context;
         }
 
+        #if FAKE_XRM_EASY_9
+        /// <summary>
+        /// Returns a fake plugin execution context for xMultiple messages from a default plugin context in code
+        /// </summary>
+        /// <param name="ctx">The fake plugin context with support for xMultiple messages</param>
+        /// <returns></returns>
+        protected IPluginExecutionContext4 GetFakedPluginContext4(XrmFakedPluginExecutionContext4 ctx)
+        {
+            var context = A.Fake<IPluginExecutionContext4>();
+
+            PopulatePluginExecutionContextPropertiesFromFakedContext(context, ctx);
+
+            A.CallTo(() => context.PreEntityImagesCollection).ReturnsLazily(() => ctx.PreEntityImagesCollection);
+            A.CallTo(() => context.PostEntityImagesCollection).ReturnsLazily(() => ctx.PostEntityImagesCollection);
+
+            return context;
+        }
+        #endif
         
         /// <summary>
         /// Returns a fake execution context
@@ -231,6 +256,19 @@ namespace FakeXrmEasy.Plugins
                     A.CallTo(() => context.PrimaryEntityName).ReturnsLazily(() => target.LogicalName);
                 }
             }
+        }
+
+        /// <summary>
+        /// Populates IPluginExecutionContext properties from a XrmFakedPluginExecutionContext
+        /// </summary>
+        /// <param name="context"></param>
+        /// <param name="ctx"></param>
+        protected void PopulatePluginExecutionContextPropertiesFromFakedContext(IPluginExecutionContext context,
+            XrmFakedPluginExecutionContext ctx)
+        {
+            PopulateExecutionContextPropertiesFromFakedContext(context, ctx);
+            A.CallTo(() => context.ParentContext).ReturnsLazily(() => ctx.ParentContext);
+            A.CallTo(() => context.Stage).ReturnsLazily(() => ctx.Stage);
         }
     }
 }